Fix the disc number migration in case there are no obsolete albums
There was an error in the DiskNumberMigration logic, which broke the SW
update process in case when no album entities had to bee merged and
hence, the list of obsolete albums to remove was empty.

This error was introduced in the commit 4a468d327.

+++ migration/disknumbermigration.php
@@ -129,10 +129,15 @@ class DiskNumberMigration implements IRepairStep {
 	 * Delete from the albums table those rows which were made obsolete by the previous steps
 	 */
 	private function removeObsoleteAlbums() {
-		$sql = 'DELETE FROM `*PREFIX*music_albums` '.
-				'WHERE `id` IN '. $this->questionMarks(count($this->obsoleteAlbums));
+		$count = count($this->obsoleteAlbums);
 
-		return $this->db->executeUpdate($sql, $this->obsoleteAlbums);
+		if ($count > 0) {
+			$sql = 'DELETE FROM `*PREFIX*music_albums` '.
+					'WHERE `id` IN '. $this->questionMarks($count);
+			$count = $this->db->executeUpdate($sql, $this->obsoleteAlbums);
+		}
+
+		return $count;
 	}
 
 	/**
